+++ packages/jest-dom-mocks/CHANGELOG.md
@@ -7,6 +7,18 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [2.9.3] - 2020-10-20
+
+- Updated `tslib` dependency to `^1.14.1`. [#1657](https://github.com/Shopify/quilt/pull/1657)
+
+## [2.9.2] - 2020-10-09
+
+### Added
+
+- `fetch-mock` is updated to the latest version [#1510](https://github.com/Shopify/quilt/pull/1510)
+
+  Please see the [migration guide](./migration-guide.md) for more information.
+
 ## [2.9.1] - 2020-05-05
 
 ### Added