+++ packages/react-form-state/CHANGELOG.md
@@ -7,6 +7,16 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [0.11.28] - 2020-10-23
+
+### Fixed
+
+- Export `StringMapper` interface [#1655](https://github.com/Shopify/quilt/pull/1655)
+
+## [0.11.27] - 2020-10-20
+
+- Updated `tslib` dependency to `^1.14.1`. [#1657](https://github.com/Shopify/quilt/pull/1657)
+
 ## [0.11.11] - 2019-10-01
 
 ### Added