+++ packages/react-import-remote/CHANGELOG.md
@@ -7,7 +7,17 @@ and adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
-## [2.0.0] - 2019-05-03
+## [1.1.3] - 2020-10-20
+
+- Updated `tslib` dependency to `^1.14.1`. [#1657](https://github.com/Shopify/quilt/pull/1657)
+
+## [1.1.0] - 2020-09-15
+
+### Fixed
+
+- Exported `useImportRemote` hook and `Status`([1634](https://github.com/Shopify/quilt/pull/1634))
+
+## [1.0.0] - 2019-05-03
 
 ### Changed